requirements.txt
Flask
datetime
uuid
Flask-SQLAlchemy
PyJWT
bs4
python-dateutil
psycopg2
playwright
